# api-rest-amazonia/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/organizaciones": {
      "get": {
        "description": "Token opcional. Sin token → devuelve datos resumidos (card). Con token válido → devuelve objeto completo.",
        "operationId": "OrganizacionesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de la página (empieza en 1). Si no se envía, se toma por defecto el valor 1.",
            "schema": {
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Cantidad de resultados por página. Si no se envía, se toma por defecto el valor 10.",
            "schema": {
              "maximum": 100,
              "default": 10,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "departamento",
            "required": false,
            "in": "query",
            "description": "Filtrar por ID de departamento",
            "schema": {
              "example": 3,
              "type": "number"
            }
          },
          {
            "name": "esNacional",
            "required": false,
            "in": "query",
            "description": "Filtrar por organizaciones nacionales (true) o locales (false)",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "tipo",
            "required": false,
            "in": "query",
            "description": "Filtrar por ID de tipo de organización",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Buscar por nombre de organización (búsqueda parcial, insensible a mayúsculas)",
            "schema": {
              "example": "Fundación",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado paginado. Shape varía según autenticación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar organizaciones con filtros y paginación",
        "tags": [
          "Organizaciones",
          "Organizaciones — Público"
        ]
      }
    },
    "/api/organizaciones/{id}": {
      "get": {
        "operationId": "OrganizacionesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organización con tipo, departamento, proyectos y empresas relacionadas"
          },
          "401": {
            "description": "Sin token o token inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🔒 Detalle completo de una organización. Requiere token JWT.",
        "tags": [
          "Organizaciones",
          "Organizaciones — Autenticado"
        ]
      }
    },
    "/api/organizaciones/{id}/logo": {
      "post": {
        "operationId": "OrganizacionesController_uploadLogo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ logoUrl: string }"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🔒 Subir o reemplazar logo de organización (máx 2 MB, JPEG/PNG/WebP → WebP 400×400)",
        "tags": [
          "Organizaciones",
          "Organizaciones — Autenticado"
        ]
      },
      "delete": {
        "operationId": "OrganizacionesController_deleteLogo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Logo eliminado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🔒 Eliminar logo de organización",
        "tags": [
          "Organizaciones",
          "Organizaciones — Autenticado"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kaa Iya — Backend API",
    "description": "API REST de la plataforma **Kaa Iya** (\"Espíritu del Bosque\"), para gestionar y visibilizar iniciativas sostenibles en la Amazonía boliviana.\n\n**Base URL:** `/api`\n\n**Autenticación:** Bearer JWT. Obtén tu token en `POST /auth/login` e inclúyelo en todas las peticiones protegidas: `Authorization: Bearer <token>`",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo Kaa Iya — UCB San Pablo",
      "url": "https://www.ucb.edu.bo",
      "email": "catedra.amazonia@ucb.edu.bo"
    },
    "license": {
      "name": "UNLICENSED",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development"
    },
    {
      "url": "https://kaaiya-backend.onrender.com",
      "description": "Producción (Render/Railway — ajustar al dominio real)"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Ingresa el token JWT obtenido de POST /auth/login"
      }
    },
    "schemas": {
...
      "PaginationResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Array de elementos de la página actual",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "page": {
            "type": "number",
            "description": "Número de la página actual",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Cantidad de elementos por página",
            "example": 10
          },
          "pages": {
            "type": "number",
            "description": "Cantidad total de páginas disponibles",
            "example": 5
          },
          "total": {
            "type": "number",
            "description": "Cantidad total de registros encontrados",
            "example": 42
          },
          "has_next": {
            "type": "boolean",
            "description": "Indica si existe una página siguiente",
            "example": true
          },
          "has_prev": {
            "type": "boolean",
            "description": "Indica si existe una página anterior",
            "example": false
          }
        },
        "required": [
          "data",
          "page",
          "limit",
          "pages",
          "total",
          "has_next",
          "has_prev"
        ]
      },
...
    }
  }
}